Add class for inline icons
This simple class handles alignment for svg and img icons inline with
text. A common problem is getting icons to visually align with text.
This icon inherits size from its parent's font-size, and adjusts its
position to be visually centered. This alignment can be modified
depending on the properties of the font it's combined with.