Unify linter versions, ignore from Dependabot
This locks ESLint and stylelint to specific versions and ignores them
from Dependabot so that it doesn't automatically open Pull Requests
to update them.

The reason is because Hound has specific versions of ESLint and
stylelint that it runs, and so while we might be using v12 of stylelint
locally, for example, Hound is running v10 in our Pull Requests and so
the results can differ.

Hound publishes their [supported linter versions][supported-linters]
and we can configure Hound to use specific versions via its `.hound.yml`
config.

Since these are just development tools, I think its probably fine for
them to be updated manually and less frequent than with Dependabot.

[supported-linters]: http://help.houndci.com/en/articles/2461415-supported-linters